T#776: Extract specs (Spec Review SDD workflow) to src/specs/routes.ts
Phase 1.11 of Backend Modularization (Library #102). Mechanical move, no
logic changes.

**Moved to src/specs/routes.ts (~641L)**:
- CREATE TABLE: spec_reviews, spec_versions, spec_comments (idempotent)
- parseVersionN helper
- 19 spec routes total:
  - GET /api/specs (list)
  - GET /api/specs/:id, GET /api/specs/:id/content, /:id/versions, /:id/history, /:id/diff
  - POST /api/specs (submit), POST /api/specs/:id/review (Tier 1/2/3 pen-stamps)
  - GET /api/specs/:id/links, POST /api/specs/:id/link, DELETE /api/specs/:id/link
  - GET /api/specs/by-task/:taskId, GET /api/specs/by-thread/:threadId
  - POST /api/specs/:id/resubmit, POST /api/specs/:id/reopen
  - DELETE /api/specs/:id
  - Spec Comments (T#332): GET /api/specs/:id/comments, GET /api/spec-comments/:commentId, POST /api/specs/:id/comments

**Exports**:
- registerSpecsRoutes(app, sqlite, { hasSessionAuth, requireBeastIdentity, isTrustedRequest, wsBroadcast })

**Imports**:
- searchIndexUpsert, searchIndexDelete from ../search/routes.ts (cross-domain, specs write paths re-index for search)
- addMessage from ../forum/handler.ts (spec→forum notification thread on submit/review)
- ../forum/mentions.ts dynamic-import for @-mention parsing in spec comments (path corrected ./forum → ../forum)

**Server.ts changes**:
- Added import after integrations import
- Removed L6176-6817 (specs block, ~642 lines) — replaced with 4-line register call
- Net: server.ts down ~637 lines (was 8,966, now 8,329)

**Verification**:
- Boot test: server starts clean. All 19 spec routes register cleanly.
- tsc --noEmit: zero new errors in src/specs/routes.ts
- No logic changes: all SDD-workflow gates preserved (Tier 1/2/3 reviewer
  validation, version chain via parseVersionN, resubmit/reopen state transitions,
  T#332 spec comments with mentions)

**Cumulative across T#771-T#776**: server.ts -3,634 lines (was 11,963, now 8,329) — 30% reduction
